server/scripts/requirements.txt
ebooklib==0.18
pymongo==4.6.1
python-dotenv==1.0.1
Pillow==10.2.0
beautifulsoup4==4.12.3
lxml==5.1.0 